refactor(chat): メッセージパートとテキストコンテンツのコンポーネントに分離し、インポートを整理
